requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1        
aiosignal==1.2.0      
anyio==3.5.0
Async-OAuthlib==0.0.9 
async-timeout==4.0.2  
attrs==21.4.0
base91==1.0.1
beautifulsoup4==4.10.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
chess==1.9.0
click==8.1.2
colorama==0.4.4
cssselect2==0.5.0
Flask==2.1.1
frozenlist==1.3.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.12.3
httpx==0.16.1
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imgurpython==1.1.7
instant-api-client==0.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
multidict==6.0.2
nextcord==2.0.0a10
nextcord-ext-activities==2022.3.14
numpy==1.22.3
oauthlib==3.2.0
Pillow==9.1.0
praw==7.5.0
prawcore==2.3.0
priority==2.0.0
prsaw==0.4.0
pyfiglet==0.8.post1
PyJWT==2.3.0
python-dotenv==0.20.0
PyYAML==6.0
quart==0.17.0
Quart-Discord==2.1.4
reportlab==3.6.9
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.3.1
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
typing_extensions==4.1.1
update-checker==0.18.0
urbandictionary==1.1
urllib3==1.26.9
wavelink==1.2.4
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
wikipedia==1.4.0
Wikipedia-API==0.5.4
wsproto==1.1.0
yarl==1.7.2
